Make `CompletionOperation`s private
`CompletionOperation` is only used internally by the queues, and the
reason why it was `public` was for it to be available in the Obj-C
queue.

To avoid exposing it simply because of this, `CompletionOperation` was
marked `private`, and a `APNCompletionOperation` Obj-C version was
created in `APNSubGroupOperationQueue.m`.

A debug `NSLog` was also removed from the ObjC tests.